requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Booktype==1.5
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
kombu==5.4.2
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
